Add ROM_bank_to_be_mapped_on_slot1/0 symbols
ROM_bank_to_be_mapped_on_slot2 already exists and can be written
to directly to switch bank, and read to know which bank is active.

When using banked code (in slot 1) there are cases where one needs to
know which code bank is active, and reading from FFFE is one way to
do this.

I'm not sure what the use cases for slot0 are, but defining a symbol
to access it should not hurt.