Add support for splitting assets over multiple banks.
If --allowsplitting is passed on the command-line, when an asset
larger than 16kB is encountered, instead of generating an error,
assets2banks will split it in multiple 16kB chunks and place
those in consecutive banks.

The string "_PARTx" where x is the part number (starting at 0)
will be appended to the default resource name.

For instance, bigfile would be declared as

bigfile_PART0
bigfile_PART1