Use the A register explicitely with OR instruction
"OR L" and "OR A, L" are equivalent, but when compiling with --peep-asm,
the optimiser does not seem to know that OR L touches A, so it deems
the "LD A,H" instruction preceding useless and drops it.

i.e. with sdcc 4.2.8 #13663, the following:

ld hl,(_SMS_theFrameInterruptHandler)
ld a,h
or l

Becomes:

ld  hl,(_SMS_theFrameInterruptHandler)
or  l

As the A register is often non-zero, the frame interrupt handler function ends
up being called, even when _SMS_theFrameInterruptHandler is NULL.