Fix the putchar() function
In SMSlib, putchar() was implemented as void putchar(char). However,
SDCC defines putchar() as extern "int putchar(int)" in stdio.h, and
this is where user code gets the function prototype from.

With the new ABI this is a problem, since the caller puts the character
value in HL because it is 16 bits, but compiled code in SMSlib expects
it in A since it is 8 bit...

With this patch it is now working again.

I include <stdio.h> so the compiler will issue a warning or
error if the definition in stdio.h ever changes, though it is not
likely...