Workaround compilation bug - likely faster implementation now too
SDCC 4.2.0 apparently does not compile SMS_loadTileMapArea correctly.
When building the lib with this version of SDCC, it appears to copy a
wider than expecter area, but the generated asm is too confusing for
me to understanding it, track down the problem and create a simple
test case to submit a bug...

Suspecting it may have had something to do with the presence of an inline
function, so I tried inlining it manually - but ended up changing how
things are done quite a bit.

Though I do not have measurements, I think this, while a lot less compact
in terms of code lines, is a faster implementation since the calculation
of the target VRAM address is done only once outside the loop now.
(I confirmed by comparing the assembly generated before and after my changes)